Updated taunt and taunt.spec to include recent changes.
Recent changes to the code needed both taunt and taunt.spec to update in order
to pass the new tests. This mostly had to do with the fact that registration is
no longer required (the commit that closed issue #89). Changes included changing
"client.owner" and "client.owner.game" to "client".