Change Default Leaderboard Max
We were constantly typing in 'leaderboar 15' to see everyone on the list. Making the default larger made it feel more natural.  If you want to see the top three or five or ten, you'd have to say that.  Past 15, we don't really care too much so 15 seemed like a good default.